requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.6.16
decorator==4.4.0
elasticsearch==7.0.2
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
numpy==1.17.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
